+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/Actor/ActorDto.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Collections.Generic;
+using Energinet.DataHub.MarketParticipant.Domain.Model;
 
 namespace Energinet.DataHub.MarketParticipant.Application.Commands.Actor
 {
@@ -21,5 +22,6 @@ namespace Energinet.DataHub.MarketParticipant.Application.Commands.Actor
         string ExternalActorId,
         GlobalLocationNumberDto Gln,
         string Status,
-        IEnumerable<MarketRoleDto> MarketRoles);
+        IEnumerable<MarketRoleDto> MarketRoles,
+        IEnumerable<MeteringPointType> MeteringPointTypes);
 }