+++ source/Energinet.DataHub.MarketParticipant.Domain/Model/GridAreaName.cs
@@ -12,16 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MarketParticipant.Utilities
+namespace Energinet.DataHub.MarketParticipant.Domain.Model
 {
-    public static class Guard
-    {
-        public static void ThrowIfNull(object element, string name)
-        {
-            if (element is null)
-            {
-                throw new System.ArgumentNullException(name);
-            }
-        }
-    }
+    public sealed record GridAreaName(string Value);
 }