+++ source/Energinet.DataHub.MarketParticipant.Infrastructure/Persistence/EntityConfiguration/GridAreaEntityConfiguration.cs
@@ -14,7 +14,6 @@
 
 using Energinet.DataHub.MarketParticipant.Domain.Model;
 using Energinet.DataHub.MarketParticipant.Infrastructure.Persistence.Model;
-using Energinet.DataHub.MarketParticipant.Utilities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
@@ -24,13 +23,14 @@ namespace Energinet.DataHub.MarketParticipant.Infrastructure.Persistence.EntityC
     {
         public void Configure(EntityTypeBuilder<GridAreaEntity> builder)
         {
-            Guard.ThrowIfNull(builder, nameof(builder));
+            ArgumentNullException.ThrowIfNull(builder, nameof(builder));
             builder.ToTable("GridAreaNew");
             builder.HasKey(gridArea => gridArea.Id);
             builder.Property(p => p.Id).ValueGeneratedOnAdd();
-            builder.Property(p => p.PriceAreaCode).HasConversion(
-                to => to.Name,
-                from => PriceAreaCode.FromName(from, true));
+            builder.Property(p => p.PriceAreaCode)
+                .HasConversion(
+                    to => to.Name,
+                    from => PriceAreaCode.FromName(from, true));
         }
     }
 }