+++ source/Energinet.DataHub.MarketParticipant.Integration.Model/source/Energinet.DataHub.MarketParticipant.Integration.Model/Parsers/GridAreaUpdatedIntegrationEventParser.cs
@@ -20,33 +20,13 @@ using Google.Protobuf;
 
 namespace Energinet.DataHub.MarketParticipant.Integration.Model.Parsers
 {
-    internal sealed class GridAreaUpdatedIntegrationEventParser : IGridAreaUpdatedIntegrationEventParser
+    public sealed class GridAreaUpdatedIntegrationEventParser : IGridAreaUpdatedIntegrationEventParser
     {
-        public GridAreaUpdatedIntegrationEvent Parse(byte[] protoContract)
-        {
-            try
-            {
-                var contract = GridAreaUpdatedIntegrationEventContract.Parser.ParseFrom(protoContract);
-
-                return new GridAreaUpdatedIntegrationEvent(
-                    Guid.Parse(contract.Id),
-                    Guid.Parse(contract.GridAreaId),
-                    contract.Name,
-                    contract.Code,
-                    Enum.IsDefined((PriceAreaCode)contract.PriceAreaCode) ? (PriceAreaCode)contract.PriceAreaCode : throw new FormatException(nameof(contract.PriceAreaCode)),
-                    Guid.Parse(contract.GridAreaLinkId));
-            }
-            catch (Exception ex) when (ex is InvalidProtocolBufferException or FormatException)
-            {
-                throw new MarketParticipantException($"Error parsing byte array for {nameof(GridAreaUpdatedIntegrationEvent)}", ex);
-            }
-        }
-
         public byte[] Parse(GridAreaUpdatedIntegrationEvent integrationEvent)
         {
             try
             {
-                Guard.ThrowIfNull(integrationEvent, nameof(integrationEvent));
+                ArgumentNullException.ThrowIfNull(integrationEvent, nameof(integrationEvent));
 
                 var contract = new GridAreaUpdatedIntegrationEventContract
                 {
@@ -63,6 +43,26 @@ namespace Energinet.DataHub.MarketParticipant.Integration.Model.Parsers
             catch (Exception ex) when (ex is InvalidProtocolBufferException)
             {
                 throw new MarketParticipantException($"Error parsing {nameof(GridAreaUpdatedIntegrationEvent)}", ex);
+            }
+        }
+
+        internal GridAreaUpdatedIntegrationEvent Parse(byte[] protoContract)
+        {
+            try
+            {
+                var contract = GridAreaUpdatedIntegrationEventContract.Parser.ParseFrom(protoContract);
+
+                return new GridAreaUpdatedIntegrationEvent(
+                    Guid.Parse(contract.Id),
+                    Guid.Parse(contract.GridAreaId),
+                    contract.Name,
+                    contract.Code,
+                    Enum.IsDefined((PriceAreaCode)contract.PriceAreaCode) ? (PriceAreaCode)contract.PriceAreaCode : throw new FormatException(nameof(contract.PriceAreaCode)),
+                    Guid.Parse(contract.GridAreaLinkId));
+            }
+            catch (Exception ex) when (ex is InvalidProtocolBufferException or FormatException)
+            {
+                throw new MarketParticipantException($"Error parsing byte array for {nameof(GridAreaUpdatedIntegrationEvent)}", ex);
             }
         }
     }