+++ source/Energinet.DataHub.MarketParticipant.Integration.Model/documents/release-notes/release-notes.md
@@ -1,9 +1,5 @@
 # Energinet.DataHub.MarketParticipant.Integration.Model Release notes
 
-## Version 2.0.0
-
-- .NET 6 upgrade
-
 ## Version 1.1.0
 
 - Preparing package for release.
@@ -20,3 +16,19 @@
 
 - Adding shared contract parser returning base integration event.
 - Use of this new parser is required.
+
+## Version 2.0.0
+
+- .NET 6 upgrade
+
+## Version 2.0.1
+
+- Public parse methods for helping with integration testing.
+
+## Version 2.0.3
+
+- Added metering point types.
+
+## Version 2.0.4
+
+- Address fields can no longer be null, but may be empty.